Renamed test_data.py to query_data.py
