+++ test/test_query.py
@@ -1,4 +1,4 @@
-from test.data.test_data import (
+from test.data.query_data import (
     get_blood_and_immune_test_data,
     get_context_list,
     get_context_members_result,
